requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
anthropic>=0.7.8
google-generativeai>=0.3.2
opencv-python>=4.8.1.78
pillow>=10.3.0
numpy>=1.26.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
pypdf2>=3.0.1
pdf2image>=1.16.3

